removed prefix store from store package
